% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_year_stats.R
\name{add_comparison_stats}
\alias{add_comparison_stats}
\title{Add Comparison Stats}
\usage{
add_comparison_stats(player_df_initial, comparison_date)
}
\arguments{
\item{comparison_date}{}
}
\description{
Add Comparison Stats
}
