% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_year_stats.R
\name{add_current_season_reduced}
\alias{add_current_season_reduced}
\title{add_current_season_reduced}
\usage{
add_current_season_reduced(data_initial)
}
\arguments{
\item{data_initial}{}
}
\description{
add_current_season_reduced
}
