% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{coalesce_join}
\alias{coalesce_join}
\title{coalesce_join}
\usage{
coalesce_join(x, y, by = NULL, suffix = c(".x", ".y"),
  join = dplyr::full_join, ...)
}
\arguments{
\item{data_initial}{}
}
\description{
coalesce_join
}
