% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs_ids.R
\name{fangraph_name_changes}
\alias{fangraph_name_changes}
\title{fangraph_name_changes}
\usage{
fangraph_name_changes(initial_data)
}
\arguments{
\item{initial_data}{}
}
\description{
fangraph_name_changes
}
