% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_excel.R
\name{load_salaries}
\alias{load_salaries}
\title{Load player salaries from LWNNsalaries.xls for each team}
\usage{
load_salaries()
}
\value{
DF with salaries for each player
}
\description{
Load player salaries from LWNNsalaries.xls for each team
}
