% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_lwnn_unsigned}
\alias{output_lwnn_unsigned}
\title{Output LWNN Unsigned}
\usage{
output_lwnn_unsigned(file_name, comparison_date)
}
\arguments{
\item{comparison_date}{}
}
\description{
Output LWNN Unsigned
}
