% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs_ids.R
\name{prospect_name_updates}
\alias{prospect_name_updates}
\title{prospect_name_updates}
\usage{
prospect_name_updates(data_initial)
}
\arguments{
\item{data_initial}{}
}
\description{
prospect_name_updates
}
